RT#7869: fix error messages when send_email fails in the job queue

+++ FS/FS/Daemon.pm
@@ -83,6 +83,7 @@ sub myexit {
 }
 
 sub _die {
+  die @_ if $^S; # $^S = 1 during an eval(), don't break exception handling
   my $msg = shift;
   unlink $pid_file if -e $pid_file;
   _logmsg($msg);
